a lot of things have changed
Some cleanup has been done since I did not follow the planned architecture to the expected
quality. This mainly refers to the app and the importer but also a bit for the egui app.
There I did some work to create tabs and just a better interface.

# Egui App

The main thing here is that I created a Grid for the images with pagination so that only a
few of the images have to be loaded at a time as well as some controls for the pagination.

# App Module

Heic improvements are the most interesting part here. It seems that resizing and saving as jpeg
works fine. Other thing is also file pathing.